+++ src/Game.php
@@ -1,0 +1,14 @@
+<?php namespace JoeStrong\RetroAchievements;
+
+/**
+ * Represents a game on RetroAchievements.org
+ *
+ * @package JoeStrong\RetroAchievements
+ */
+class Game
+{
+    public $id;
+    public $title;
+    public $consoleId;
+    public $imageIcon;
+}