+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## dev-develop
 
 - Added \RetroAchievements\getConsoles()
+- Added \RetroAchievements\getGamesForConsole($consoleId)
 - Added PHP-CS-Fixer for PSR-2 formatting
 
 ## 0.1.0